+++ examples/gltf/main.rs
@@ -116,7 +116,7 @@ impl<'a, 'b> SimpleState<'a, 'b> for Example {
     fn handle_event(
         &mut self,
         data: StateData<GameData>,
-        event: StateEvent<()>,
+        event: StateEvent,
     ) -> SimpleTrans<'a, 'b> {
         let StateData { world, .. } = data;
         if let StateEvent::Window(event) = &event {