+++ src/app.rs
@@ -5,9 +5,9 @@ use assets::{Loader, Source};
 use core::frame_limiter::{FrameLimiter, FrameRateLimitConfig, FrameRateLimitStrategy};
 use core::shrev::{EventChannel, ReaderId};
 use core::timing::{Stopwatch, Time};
-use core::Named;
+use core::{EventReader, Named};
 use ecs::common::Errors;
-use ecs::prelude::{Component, World};
+use ecs::prelude::{Component, World, Write, Read};
 use error::{Error, Result};
 use game_data::DataInit;
 use log::Level;
@@ -15,6 +15,7 @@ use rayon::ThreadPoolBuilder;
 use shred::Resource;
 use state::{State, StateData, StateMachine};
 use state_event::StateEvent;
+use state_event::StateEventReader;
 use std::error::Error as StdError;
 use std::marker::PhantomData;
 use std::path::Path;
@@ -22,7 +23,34 @@ use std::sync::Arc;
 use std::time::Duration;
 #[cfg(feature = "profiler")]
 use thread_profiler::{register_thread_with_profiler, write_profile};
-use winit::{Event, WindowEvent};
+use winit::Event;
+
+/// `CoreApplication` is the application implementation for the game engine. This is fully generic
+/// over the state type and event type.
+///
+/// When starting out with Amethyst, use the type alias `Application`, which have sensible defaults
+/// for the `Event` and `EventReader` generic types.
+///
+/// ### Type parameters:
+///
+/// - `T`: `State`
+/// - `E`: `Event` type that should be sent to the states
+/// - `R`: `EventReader` implementation for the given event type `E`
+#[derive(Derivative)]
+#[derivative(Debug)]
+pub struct CoreApplication<'a, T, E = StateEvent, R = StateEventReader> {
+    /// The world
+    #[derivative(Debug = "ignore")]
+    world: World,
+    #[derivative(Debug = "ignore")]
+    reader: R,
+    #[derivative(Debug = "ignore")]
+    events: Vec<E>,
+    event_reader_id: ReaderId<Event>,
+    states: StateMachine<'a, T, E>,
+    ignore_window_close: bool,
+    data: T,
+}
 
 /// An Application is the root object of the game engine. It binds the OS
 /// event loop, state machines, timers and other core components in a central place.
@@ -93,21 +121,12 @@ use winit::{Event, WindowEvent};
 /// ```
 ///
 /// [log]: https://crates.io/crates/log
-#[derive(Derivative)]
-#[derivative(Debug)]
-pub struct Application<'a, T, E: Send + Sync + 'static> {
-    /// The world
-    #[derivative(Debug = "ignore")]
-    world: World,
-    window_reader_id: ReaderId<Event>,
-    ui_reader_id: ReaderId<UiEvent>,
-    custom_reader_id: ReaderId<E>,
-    states: StateMachine<'a, T, E>,
-    ignore_window_close: bool,
-    data: T,
-}
+pub type Application<'a, T> = CoreApplication<'a, T, StateEvent, StateEventReader>;
 
-impl<'a, T, E: Send + Sync + Clone + 'static> Application<'a, T, E> {
+impl<'a, T, E, R> CoreApplication<'a, T, E, R>
+where
+    E: Clone + Send + Sync + 'static,
+{
     /// Creates a new Application with the given initial game state.
     /// This will create and allocate all the needed resources for
     /// the event loop of the game engine. It is a shortcut for convenience
@@ -161,6 +180,8 @@ impl<'a, T, E: Send + Sync + Clone + 'static> Application<'a, T, E> {
         P: AsRef<Path>,
         S: State<T, E> + 'a,
         I: DataInit<T>,
+        for<'b> R: EventReader<'b, Event = E>,
+        R: Default,
     {
         ApplicationBuilder::new(path, initial_state)?.build(init)
     }
@@ -169,10 +190,11 @@ impl<'a, T, E: Send + Sync + Clone + 'static> Application<'a, T, E> {
     ///
     /// This is identical in function to
     /// [ApplicationBuilder::new](struct.ApplicationBuilder.html#method.new).
-    pub fn build<P, S>(path: P, initial_state: S) -> Result<ApplicationBuilder<S, E>>
+    pub fn build<P, S>(path: P, initial_state: S) -> Result<ApplicationBuilder<S, E, R>>
     where
         P: AsRef<Path>,
-        S: State<T, E>,
+        S: State<T, E> + 'a,
+        for<'b> R: EventReader<'b, Event = E>,
     {
         ApplicationBuilder::new(path, initial_state)
     }
@@ -186,7 +208,10 @@ impl<'a, T, E: Send + Sync + Clone + 'static> Application<'a, T, E> {
     ///
     /// See the example supplied in the
     /// [`new`](struct.Application.html#examples) method.
-    pub fn run(&mut self) {
+    pub fn run(&mut self)
+    where
+        for<'b> R: EventReader<'b, Event = E>,
+    {
         self.initialize();
         self.world.write_resource::<Stopwatch>().start();
         while self.states.is_running() {
@@ -216,64 +241,62 @@ impl<'a, T, E: Send + Sync + Clone + 'static> Application<'a, T, E> {
             .expect("Tried to start state machine without any states present");
     }
 
-    /// Advances the game world by one tick.
-    fn advance_frame(&mut self) {
-        trace!("Advancing frame (`Application::advance_frame`)");
+    // React to window close events
+    fn should_close(&mut self) -> bool {
+        if self.ignore_window_close {
+            false
+        } else {
+            use renderer::WindowEvent;
+            let world = &mut self.world;
+            let reader_id = &mut self.event_reader_id;
+            world.exec(|ev : Read<EventChannel<Event>>| {
+                ev.read(reader_id).any(|e| {
+                    if cfg!(target_os = "ios") {
+                        if let Event::WindowEvent {event: WindowEvent::Destroyed,  .. } = e {
+                            true
+                        } else {
+                            false
+                        }
+                    } else {
+                        if let Event::WindowEvent {event: WindowEvent::CloseRequested,  .. } = e {
+                            true
+                        } else {
+                            false
+                        }
+                    }
+                })
+            })
+        }
+    }
 
-        {
+    /// Advances the game world by one tick.
+    fn advance_frame(&mut self)
+    where
+        for<'b> R: EventReader<'b, Event = E>,
+    {
+        trace!("Advancing frame (`Application::advance_frame`)");
+        if self.should_close() {
             let world = &mut self.world;
             let states = &mut self.states;
+            states.stop(StateData::new(world, &mut self.data));
+        }
+
+
+        {
             #[cfg(feature = "profiler")]
             profile_scope!("handle_event");
 
-            let window_events = world
-                .read_resource::<EventChannel<Event>>()
-                .read(&mut self.window_reader_id)
-                .cloned()
-                .map(|e| StateEvent::Window(e))
-                .collect::<Vec<_>>();
+            {
+                let events = &mut self.events;
+                self.reader.read(self.world.system_data(), events);
+            }
 
-            for event in window_events {
-                if !self.ignore_window_close {
-                    if cfg!(target_os = "ios") {
-                        if let &StateEvent::Window(Event::WindowEvent {
-                            event: WindowEvent::Destroyed,
-                            ..
-                        }) = &event
-                        {
-                            states.stop(StateData::new(world, &mut self.data));
-                        }
-                    } else {
-                        if let &StateEvent::Window(Event::WindowEvent {
-                            event: WindowEvent::CloseRequested,
-                            ..
-                        }) = &event
-                        {
-                            states.stop(StateData::new(world, &mut self.data));
-                        }
-                    }
+            {
+                let world = &mut self.world;
+                let states = &mut self.states;
+                for e in self.events.drain(..) {
+                    states.handle_event(StateData::new(world, &mut self.data), e);
                 }
-                states.handle_event(StateData::new(world, &mut self.data), event);
-            }
-
-            let ui_events = world
-                .read_resource::<EventChannel<UiEvent>>()
-                .read(&mut self.ui_reader_id)
-                .cloned()
-                .map(|e| StateEvent::Ui(e))
-                .collect::<Vec<_>>();
-            for event in ui_events {
-                states.handle_event(StateData::new(world, &mut self.data), event);
-            }
-
-            let custom_events = world
-                .read_resource::<EventChannel<E>>()
-                .read(&mut self.custom_reader_id)
-                .cloned()
-                .map(|e| StateEvent::Custom(e))
-                .collect::<Vec<_>>();
-            for event in custom_events {
-                states.handle_event(StateData::new(world, &mut self.data), event);
             }
         }
         {
@@ -314,7 +337,7 @@ impl<'a, T, E: Send + Sync + Clone + 'static> Application<'a, T, E> {
 }
 
 #[cfg(feature = "profiler")]
-impl<'a, T, E: Send + Sync + 'static> Drop for Application<'a, T, E> {
+impl<'a, T, E, R> Drop for CoreApplication<'a, T, E, R> {
     fn drop(&mut self) {
         // TODO: Specify filename in config.
         use utils::application_root_dir;
@@ -328,16 +351,16 @@ impl<'a, T, E: Send + Sync + 'static> Drop for Application<'a, T, E> {
 /// using a custom set of configuration. This is the normal way an
 /// [`Application`](struct.Application.html)
 /// object is created.
-pub struct ApplicationBuilder<S, E> {
+pub struct ApplicationBuilder<S, E, R> {
     // config: Config,
     initial_state: S,
     /// Used by bundles to access the world directly
     pub world: World,
     ignore_window_close: bool,
-    phantom: PhantomData<E>,
+    phantom: PhantomData<(E, R)>,
 }
 
-impl<S, E: Send + Sync + 'static> ApplicationBuilder<S, E> {
+impl<S, E, X> ApplicationBuilder<S, E, X> {
     /// Creates a new [ApplicationBuilder](struct.ApplicationBuilder.html) instance
     /// that wraps the initial_state. This is the more verbose way of initializing
     /// your application if you require specific configuration details to be changed
@@ -399,7 +422,7 @@ impl<S, E: Send + Sync + 'static> ApplicationBuilder<S, E> {
     /// game.run();
     /// ~~~
 
-    pub fn new<P: AsRef<Path>>(path: P, initial_state: S) -> Result<Self> {
+    pub fn new<'a, P: AsRef<Path>>(path: P, initial_state: S) -> Result<Self> {
         use rustc_version_runtime;
 
         if !log_enabled!(Level::Error) {
@@ -436,7 +459,6 @@ impl<S, E: Send + Sync + 'static> ApplicationBuilder<S, E> {
         world.add_resource(pool);
         world.add_resource(EventChannel::<Event>::with_capacity(2000));
         world.add_resource(EventChannel::<UiEvent>::with_capacity(40));
-        world.add_resource(EventChannel::<E>::with_capacity(2));
         world.add_resource(Errors::default());
         world.add_resource(FrameLimiter::default());
         world.add_resource(Stopwatch::default());
@@ -704,10 +726,13 @@ impl<S, E: Send + Sync + 'static> ApplicationBuilder<S, E> {
     ///
     /// See the [example show for `ApplicationBuilder::new()`](struct.ApplicationBuilder.html#examples)
     /// for an example on how this method is used.
-    pub fn build<'a, T, I>(mut self, init: I) -> Result<Application<'a, T, E>>
+    pub fn build<'a, T, I>(mut self, init: I) -> Result<CoreApplication<'a, T, E, X>>
     where
         S: State<T, E> + 'a,
         I: DataInit<T>,
+        E: Clone + Send + Sync + 'static,
+        X: Default,
+        for<'b> X: EventReader<'b, Event = E>,
     {
         trace!("Entering `ApplicationBuilder::build`");
 
@@ -716,29 +741,19 @@ impl<S, E: Send + Sync + 'static> ApplicationBuilder<S, E> {
         #[cfg(feature = "profiler")]
         profile_scope!("new");
 
-        let window_reader_id = self
-            .world
-            .write_resource::<EventChannel<Event>>()
-            .register_reader();
-        let ui_reader_id = self
-            .world
-            .write_resource::<EventChannel<UiEvent>>()
-            .register_reader();
-        let custom_reader_id = self
-            .world
-            .write_resource::<EventChannel<E>>()
-            .register_reader();
-
+        let mut reader = X::default();
+        reader.setup(&mut self.world.res);
         let data = init.build(&mut self.world);
+        let event_reader_id = self.world.exec(|mut ev : Write<EventChannel<Event>>| ev.register_reader());
 
-        Ok(Application {
+        Ok(CoreApplication {
             world: self.world,
             states: StateMachine::new(self.initial_state),
-            window_reader_id,
-            ui_reader_id,
-            custom_reader_id,
+            reader,
+            events: Vec::new(),
             ignore_window_close: self.ignore_window_close,
             data,
+            event_reader_id,
         })
     }
 }